--- ExamenU3.scala
@@ -1,18 +1,0 @@
-1. Importar una simple sesión Spark.
-import org.apache.spark.sql.SparkSession
-
-2. Utilice las lineas de código para minimizar errores
-import org.apache.log4j._
-Logger.getLogger("org").setLevel(Level.ERROR)
-
-3. Cree una instancia de la sesión Spark
-val spark = SparkSession.builder().getOrCreate()
-
-4. Importar la librería de Kmeans para el algoritmo de agrupamiento.
-import org.apache.spark.ml.clustering.KMeans
-
-5. Carga el dataset de Wholesale Customers Data
-val dataset = spark.read.option("header","true").option("inferSchema","true").csv("Wholesale customers data.csv")
-
-6. Seleccione las siguientes columnas: Fresh, Milk, Grocery, Frozen, Detergents_Paper, Delicassen y llamar a este conjunto feature_data
-val feature_data = dataset.select($"Fresh", $"Milk", $"Grocery", $"Frozen", $"Detergents_Paper", $"Delicassen")
